Updated some words
I changed the word `method` for `function`, I know that is kinda dumb, but when we're doing this LAB, is happening when we learn *Objects* and the students struggle a lit bit when we use the word "method" instead "function". So technically, using the word "method" for a simple function is wrong, that's why I changed the this.